Split calibration types into separate .cpp and .h files for neater code
