+++ industrial_calibration_libs/include/industrial_calibration_libs/calibration.h
@@ -24,40 +24,6 @@ struct CovarianceRequest
   std::string object_name;
 };
 
-struct CameraOnWristExtrinsicParams 
-{
-  // Known Values
-  IntrinsicsPartial intrinsics;
-  std::vector<Pose6D> base_to_tool; // For every observation.
-
-  // Unknown Values
-  Extrinsics tool_to_camera;
-  Extrinsics target_to_base;
-};
-
-struct CameraOnWristIntrinsicParams
-{
-  // Known Values
-  std::vector<Pose6D> base_to_tool; // For every observation.
-
-  // Uknown Values
-  IntrinsicsFull intrinsics;
-  Extrinsics tool_to_camera;
-  Extrinsics target_to_base;
-};
-
-
-struct CameraOnWristExtrinsicIntrinsicParams
-{
-  // Known Values
-  std::vector<Pose6D> base_to_tool; // For every observation.
-
-  // Uknown Values
-  IntrinsicsFull intrinsics;
-  Extrinsics tool_to_camera;
-  Extrinsics target_to_base;
-};
-
 class CalibrationJob
 {
 public:
@@ -95,80 +61,5 @@ protected:
 
   bool output_results_;
 };
-
-class CameraOnWristExtrinsic : public CalibrationJob
-{
-public:
-  struct Result
-  {
-    double extrinsics[6];
-    double target_to_base[6];
-  };
-
-  CameraOnWristExtrinsic(const ObservationData &observation_data,
-    const Target &target, const CameraOnWristExtrinsicParams &params);
-
-  ~CameraOnWristExtrinsic(void) { }
-
-  bool runCalibration(void);
-
-  void displayCovariance(void);
-
-  Result getResults(void) {return result_;}
-
-private:
-  double intrinsics_[4];
-  Result result_;
-};
-
-class CameraOnWristIntrinsic : public CalibrationJob
-{
-public:
-  struct Result
-  {
-    double extrinsics[6];
-    double target_to_base[6];
-    double intrinsics[9];
-  };
-
-  CameraOnWristIntrinsic(const ObservationData &observation_data,
-    const Target &target, const CameraOnWristIntrinsicParams &params);
-
-  ~CameraOnWristIntrinsic(void) { }
-
-  bool runCalibration(void);
-
-  void displayCovariance(void);
-
-  Result getResults(void) {return result_;}
-
-private:
-  Result result_;
-};
-
-class CameraOnWristExtrinsicIntrinsic : public CalibrationJob
-{
-public:
-  struct Result
-  {
-    double extrinsics[6];
-    double target_to_base[6];
-    double intrinsics[9];
-  };
-
-  CameraOnWristExtrinsicIntrinsic(const ObservationData &observation_data,
-    const Target &target, const CameraOnWristExtrinsicIntrinsicParams &params);
-
-  ~CameraOnWristExtrinsicIntrinsic(void) { }
-
-  bool runCalibration(void);
-
-  void displayCovariance(void);
-
-  Result getResults(void) {return result_;}
-
-private:
-  Result result_;
-};
 } // namespace industrial_calibration_libs
-#endif
+#endif // CALIBRATION_H
