+++ industrial_calibration_libs/include/industrial_calibration_libs/industrial_calibration_libs.h
@@ -7,6 +7,10 @@
 #include <industrial_calibration_libs/targets.h>
 #include <industrial_calibration_libs/types.h>
 
+#include <industrial_calibration_libs/calibration_types/camera_on_wrist_extrinsic.h>
+#include <industrial_calibration_libs/calibration_types/camera_on_wrist_intrinsic.h>
+#include <industrial_calibration_libs/calibration_types/camera_on_wrist_extrinsic_intrinsic.h>
+
 namespace industrial_calibration_libs
 {
